dev-requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.10
django-admin-bootstrapped==2.5.7
django-auditlog==0.4.3
django-autocomplete-light==3.2.9
django-bootstrap3==8.2.2
django-classy-tags==0.8.0
django-colorful==1.2
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-datetime-widget==0.9.3
-e /home/trjg/code/django-geo-db
django-jsonfield==1.0.1
djangorestframework==3.6.4
djangorestframework-jwt==1.10.0
googlemaps==2.5.0
gunicorn==19.7.1
mysqlclient==1.3.7
olefile==0.44
packaging==16.8
Pillow==4.1.1
PyJWT==1.5.0
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
six==1.10.0
sorl-thumbnail==12.3
whitenoise==3.3.0
